Capacity data type changed from integer to long

+++ src/main/java/com/mateuszjanczak/barrelsbeer/domain/dto/BarrelTapHitResponse.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class BarrelTapHitResponse {
     int barrelTapId;
-    int newValue;
+    long newValue;
 }
