+++ src/main/java/com/mateuszjanczak/barrelsbeer/web/rest/BarrelTapController.java
@@ -54,7 +54,7 @@ public class BarrelTapController {
     }
 
     @GetMapping(HIT_BARREL_TAP)
-    ResponseEntity<BarrelTapHitResponse> hitBarrelTap(@PathVariable int id, @PathVariable int value) {
+    ResponseEntity<BarrelTapHitResponse> hitBarrelTap(@PathVariable int id, @PathVariable long value) {
         Optional<BarrelTapHitResponse> optionalBarrelHitResponse = barrelTapService.hitBarrelTap(id, value);
         return optionalBarrelHitResponse.map(barrelTapHitResponse -> new ResponseEntity<>(barrelTapHitResponse, HttpStatus.OK)).orElseThrow(HitException::new);
     }
